server/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
PyPDF2==3.0.1
python-docx==0.8.11
pdfplumber==0.10.3
fpdf2==2.7.9
python-magic==0.4.27
pytest==7.4.0

werkzeug==3.0.0